LayoutTests/media/track/track-css-matching.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script src=../media-controls.js></script>

        <style>

        video::cue(.red, .red2) { color:red }
        #testvideo::cue(.green) { color:green }

        </style>

        <script>

        var cueNode;
        var seekedCount = 0;
        var info = [["rgb(255, 0, 0)", "rgb(0, 128, 0)", "rgb(255, 0, 0)"],
                    ["rgb(255, 255, 255)", "rgb(128, 128, 128)", "rgb(128, 128, 128)"],
                    ["rgb(255, 255, 255)", "rgb(255, 255, 255)", "rgb(128, 128, 128)"],
                    ["rgb(255, 255, 255)", "rgb(255, 255, 255)", "rgb(255, 255, 255)"]];
        
        function skipNonElements(root)
        {
            nextElementSibling = root;
            while (nextElementSibling.nodeType != 1) {
                nextElementSibling = nextElementSibling.nextSibling;
            }
            return nextElementSibling;
        }
        
        function seeked()
        {
            if (testEnded)
                return;

            if (seekedCount == 1) {
                consoleWrite("");
                consoleWrite("");
                consoleWrite("2. Test that cues are being matched properly by the ':future' pseudo class.");
            }

            cueNode = skipNonElements(textTrackDisplayElement(video, 'all-nodes').firstChild);
            skipNonElements(cueNode);
            testExpected("getComputedStyle(cueNode).color", info[seekedCount][0]);
            cueNode = skipNonElements(cueNode.nextSibling);
            testExpected("getComputedStyle(cueNode).color", info[seekedCount][1]);
            cueNode = skipNonElements(cueNode.nextSibling);
            testExpected("getComputedStyle(cueNode).color", info[seekedCount][2]);

            if (++seekedCount == info.length)
                endTest();
            else {
                consoleWrite("");
                run("video.currentTime = " + (video.currentTime + 1));
            }
        }

        function loaded()
        {
            consoleWrite("1. Test that cues are being matched properly by the class name");
            findMediaElement();
            video.src = findMediaFile('video', '../content/test');
            video.id = "testvideo";
            waitForEvent('seeked', seeked);
            waitForEvent('canplaythrough', function() { video.currentTime = 1.5; });
        }

        </script>
    </head>
    <body onload="loaded()">
        <video controls >
            <track src="captions-webvtt/styling.vtt" kind="captions" default>
        </video>
    </body>
</html>
